refactor: refactor chat component
